back-end/account_transfer_app/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
psycopg2==2.9.9
pycparser==2.22
PySocks==1.7.1
pytz==2024.1
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
django-cors-headers== 4.4.0